Fix: launch campaign modal
