--- src/components/common/Modals/ConfirmModal/index.ts
@@ -1,3 +1,0 @@
-import ConfirmModal from './ConfirmModal'
-
-export default ConfirmModal
