+++ src/components/common/Modals/LaunchCampaignModal/index.ts
@@ -1,0 +1,3 @@
+import LaunchCampaignModal from './LaunchCampaignModal'
+
+export default LaunchCampaignModal
