+++ src/components/common/Modals/index.ts
@@ -1,7 +1,7 @@
 import CreativePreviewModal from './CreativePreviewModal/CreativePreviewModal'
 import CampaignPeriodModal from './CampaignPeriodModal/CampaignPeriodModal'
 import AdminCampaignModal from './AdminCampaignModal/AdminCampaignModal'
-import ConfirmModal from './ConfirmModal/ConfirmModal'
+import LaunchCampaignModal from './LaunchCampaignModal/LaunchCampaignModal'
 import SuccessModal from './SuccessModal'
 import CustomConfirmModal from './CustomConfirmModal/CustomConfirmModal'
 
@@ -9,7 +9,7 @@ export {
   CreativePreviewModal,
   CampaignPeriodModal,
   AdminCampaignModal,
-  ConfirmModal,
+  LaunchCampaignModal,
   SuccessModal,
   CustomConfirmModal
 }
